requirements.txt
bs4>=0.0.2
chromadb>=0.6.3
dotenv>=0.9.9
google-adk>=0.3.0
google-generativeai>=0.8.5
langchain-chroma>=0.2.3
langchain-core>=0.3.56
langchain-google-genai>=2.0.10
requests>=2.32.3
fastapi>=0.115.9
uvicorn>=0.34.2